requirements.txt
inflect==5.0.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
sqlacodegen==2.3.0
SQLAlchemy==1.3.23
